Remove elf.util.set_numpy_threads
